docker/prepare_image/requirements.txt
Pillow==9.0.0
numpy